refactor: colocate publisher types
